requirements.txt
# Voice framework
pipecat-ai[google,deepgram,silero,local-smart-turn,websocket]==0.0.104

# Web framework (pipecat requires >=0.115.6)
fastapi>=0.115.6,<0.128.0
uvicorn[standard]>=0.34.0,<1

# Telephony
plivo>=4.59.0,<5

# Google Cloud TTS (gRPC streaming — pipecat[google] requires >=2.31.0)
# Already pulled in by pipecat[google], but listed explicitly for clarity
google-cloud-texttospeech>=2.31.0,<3

# Database
asyncpg>=0.30.0,<1
sqlalchemy[asyncio]>=2.0.36,<3
alembic>=1.14.0,<2

# HTTP client (for GHL API calls — pipecat requires >=3.11.12)
aiohttp>=3.11.12,<4

# Settings
pydantic-settings>=2.7.0,<3

# Logging
structlog>=24.4.0,<25
